requirements.txt
boto3
pandas
ta
logram
python-dotenv
awswrangler